небольшие улучшения wip
вынес инит стейт в отдельный модуль
