+++ src/store/initialState.js
@@ -1,0 +1,10 @@
+import films from '../mocks/films';
+const initState = {
+  genre: `All genres`,
+  films,
+  MOVIE_CARD_DEFAULT_COUNT: 20,
+  PROMO_FILM_NAME: `The Grand Budapest`,
+  PROMO_FILM_GENRE: `Drama`,
+  PROMO_FILM_RELISE_YEAR: `2020`
+};
+export default initState;
